move location to shuffle api
