+++ core/src/main/scala/org/apache/spark/TaskEndReason.scala
@@ -21,7 +21,8 @@ import java.io.{ObjectInputStream, ObjectOutputStream}
 
 import org.apache.spark.annotation.DeveloperApi
 import org.apache.spark.internal.Logging
-import org.apache.spark.scheduler.{AccumulableInfo, Location}
+import org.apache.spark.scheduler.AccumulableInfo
+import org.apache.spark.shuffle.api.Location
 import org.apache.spark.util.{AccumulatorV2, Utils}
 
 // ==============================================================================================
