+++ core/src/main/scala/org/apache/spark/shuffle/BlockStoreShuffleReader.scala
@@ -20,8 +20,8 @@ package org.apache.spark.shuffle
 import org.apache.spark._
 import org.apache.spark.internal.{config, Logging}
 import org.apache.spark.io.CompressionCodec
-import org.apache.spark.scheduler.Location
 import org.apache.spark.serializer.SerializerManager
+import org.apache.spark.shuffle.api.Location
 import org.apache.spark.storage.{BlockId, BlockManager, ShuffleBlockFetcherIterator}
 import org.apache.spark.util.CompletionIterator
 import org.apache.spark.util.collection.ExternalSorter
